@@ -31,22 +31,22 @@
     "eventsourcing"
   ],
   "devDependencies": {
-    "@types/jest": "24.0.11",
+    "@types/jest": "25.1.4",
     "@types/knex": "0.16.1",
-    "@types/node": "11.13.5",
-    "jest": "24.7.1",
+    "@types/node": "13.9.3",
+    "jest": "25.1.0",
     "npm-run-all": "4.1.5",
-    "prettier": "1.17.0",
-    "ts-jest": "24.0.2",
-    "ts-node": "8.1.0",
-    "tslint": "5.16.0",
+    "prettier": "2.0.2",
+    "ts-jest": "25.2.1",
+    "ts-node": "8.8.1",
+    "tslint": "6.1.0",
     "tslint-config-prettier": "1.18.0",
-    "typescript": "3.4.4"
+    "typescript": "3.8.3"
   },
   "dependencies": {
     "@arpinum/defender": "1.0.1",
-    "knex": "0.16.5",
-    "pg": "7.10.0",
-    "pg-query-stream": "2.0.0"
+    "knex": "0.20.13",
+    "pg": "7.18.2",
+    "pg-query-stream": "3.0.3"
   }
 }
