@@ -1,8 +1,8 @@
-'use strict';
+"use strict";
 
-const { createEventStore } = require('../build');
+const { createEventStore } = require("../build");
 
-const connectionString = 'postgres://postgres@localhost:5432/eventstoretests';
+const connectionString = "postgres://postgres@localhost:5432/eventstoretests";
 
 main();
 
@@ -11,17 +11,17 @@ async function main() {
   try {
     store = createEventStore({ connectionString });
 
-    const removeListener = store.onEvent(event =>
-      console.log('Event added:', event.type)
+    const removeListener = store.onEvent((event) =>
+      console.log("Event added:", event.type)
     );
 
-    await store.add(createEvent({ type: 'UserSignedUp' }));
-    await store.add(createEvent({ type: 'UserLoggedIn' }));
-    await store.add(createEvent({ type: 'UserLoggedOut' }));
+    await store.add(createEvent({ type: "UserSignedUp" }));
+    await store.add(createEvent({ type: "UserLoggedIn" }));
+    await store.add(createEvent({ type: "UserLoggedOut" }));
 
     removeListener();
 
-    await store.add(createEvent({ type: 'NoMoreRegistered' }));
+    await store.add(createEvent({ type: "NoMoreRegistered" }));
   } catch (e) {
     console.error(e);
     process.exit(1);
@@ -35,13 +35,13 @@ async function main() {
 function createEvent(data) {
   return Object.assign(
     {
-      type: 'UserLoggedIn',
+      type: "UserLoggedIn",
       date: new Date(),
-      targetType: 'User',
-      targetId: 'bf04b429-4c88-46de-a2ae-15624c75fd56',
+      targetType: "User",
+      targetId: "bf04b429-4c88-46de-a2ae-15624c75fd56",
       payload: {
-        login: 'johndoe'
-      }
+        login: "johndoe",
+      },
     },
     data
   );
