@@ -1,4 +1,4 @@
-import * as Knex from 'knex';
+import * as Knex from "knex";
 
 export interface Connection {
   connectionString: string;
@@ -7,8 +7,8 @@ export interface Connection {
 export function createClient(connection: Connection): Knex {
   const { connectionString } = connection;
   return Knex({
-    client: 'pg',
-    connection: connectionString
+    client: "pg",
+    connection: connectionString,
   });
 }
 
@@ -16,15 +16,15 @@ export async function createSchema(
   client: Knex,
   options?: { tableName?: string }
 ): Promise<void> {
-  const { tableName } = { tableName: 'events', ...options };
-  await client.transaction(trx =>
-    trx.schema.createTable(tableName, table => {
-      table.bigIncrements('id');
-      table.dateTime('date');
-      table.string('type').index();
-      table.string('target_type').index();
-      table.string('target_id').index();
-      table.jsonb('payload');
+  const { tableName } = { tableName: "events", ...options };
+  await client.transaction((trx) =>
+    trx.schema.createTable(tableName, (table) => {
+      table.bigIncrements("id");
+      table.dateTime("date");
+      table.string("type").index();
+      table.string("target_type").index();
+      table.string("target_id").index();
+      table.jsonb("payload");
     })
   );
 }
@@ -33,6 +33,6 @@ export async function dropSchema(
   client: Knex,
   options?: { tableName?: string }
 ): Promise<void> {
-  const { tableName } = { tableName: 'events', ...options };
-  await client.transaction(trx => trx.schema.dropTable(tableName));
+  const { tableName } = { tableName: "events", ...options };
+  await client.transaction((trx) => trx.schema.dropTable(tableName));
 }
