@@ -1,11 +1,11 @@
-import * as Knex from 'knex';
+import * as Knex from "knex";
 
-import { EventStore } from './eventStore';
-import { createDbEvent, createEvent } from './tests';
-import { OpenEventStore } from './tests/openEventStore';
-import { Event } from './types';
+import { EventStore } from "./eventStore";
+import { createDbEvent, createEvent } from "./tests";
+import { OpenEventStore } from "./tests/openEventStore";
+import { Event } from "./types";
 
-describe('Event store', () => {
+describe("Event store", () => {
   let client: Knex;
   let eventStore: OpenEventStore;
 
@@ -14,43 +14,44 @@ describe('Event store', () => {
     eventStore = new OpenEventStore({} as Knex);
   });
 
-  describe('on creation', () => {
-    it('should assert client is present', () => {
+  describe("on creation", () => {
+    it("should assert client is present", () => {
       const create = () => new EventStore(null as any);
 
-      expect(create).toThrow('client must be present');
+      expect(create).toThrow("client must be present");
     });
 
-    it('should assert options is wellformed defined', () => {
+    it("should assert options is wellformed defined", () => {
       const create = () => new EventStore(client, 3 as any);
 
-      expect(create).toThrow('options must be an object');
+      expect(create).toThrow("options must be an object");
     });
   });
 
-  describe('on add', () => {
-    it('should assert event is present', () => {
+  describe("on add", () => {
+    it("should assert event is present", () => {
       const add = eventStore.add(null as any);
 
       return add.then(
-        () => Promise.reject(new Error('Should fail')),
-        rejection => expect(rejection.message).toEqual('event must be present')
+        () => Promise.reject(new Error("Should fail")),
+        (rejection) =>
+          expect(rejection.message).toEqual("event must be present")
       );
     });
 
-    it('should assert event is wellformed', () => {
+    it("should assert event is wellformed", () => {
       const add = eventStore.add(3 as any);
 
       return add.then(
-        () => Promise.reject(new Error('Should fail')),
-        rejection =>
-          expect(rejection.message).toEqual('event must be an object')
+        () => Promise.reject(new Error("Should fail")),
+        (rejection) =>
+          expect(rejection.message).toEqual("event must be an object")
       );
     });
 
-    it('should emit added event', async () => {
+    it("should emit added event", async () => {
       const emittedEvents: Event[] = [];
-      eventStore.onEvent(event => {
+      eventStore.onEvent((event) => {
         emittedEvents.push(event);
       });
       eventStore.insertEventsInDb = () =>
@@ -62,13 +63,13 @@ describe('Event store', () => {
       expect(emittedEvents[0].id).toEqual(1);
     });
 
-    it('could stop listening to added events', async () => {
+    it("could stop listening to added events", async () => {
       const emittedEvents: Event[] = [];
-      const removeListener = eventStore.onEvent(event => {
+      const removeListener = eventStore.onEvent((event) => {
         emittedEvents.push(event);
       });
       Object.assign(eventStore, {
-        insertEventsInDb: () => Promise.resolve([createDbEvent({ id: 1 })])
+        insertEventsInDb: () => Promise.resolve([createDbEvent({ id: 1 })]),
       });
       removeListener();
 
@@ -78,44 +79,45 @@ describe('Event store', () => {
     });
   });
 
-  describe('on add all', () => {
-    it('should assert events are present', () => {
+  describe("on add all", () => {
+    it("should assert events are present", () => {
       const add = eventStore.addAll(null as any);
 
       return add.then(
-        () => Promise.reject(new Error('Should fail')),
-        rejection => expect(rejection.message).toEqual('events must be present')
+        () => Promise.reject(new Error("Should fail")),
+        (rejection) =>
+          expect(rejection.message).toEqual("events must be present")
       );
     });
 
-    it('should assert events are an array', () => {
+    it("should assert events are an array", () => {
       const add = eventStore.addAll(3 as any);
 
       return add.then(
-        () => Promise.reject(new Error('Should fail')),
-        rejection =>
-          expect(rejection.message).toEqual('events must be an array')
+        () => Promise.reject(new Error("Should fail")),
+        (rejection) =>
+          expect(rejection.message).toEqual("events must be an array")
       );
     });
 
-    it('should assert all events are wellformed', () => {
+    it("should assert all events are wellformed", () => {
       const add = eventStore.addAll([3 as any]);
 
       return add.then(
-        () => Promise.reject(new Error('Should fail')),
-        rejection =>
-          expect(rejection.message).toEqual('events[0] must be an object')
+        () => Promise.reject(new Error("Should fail")),
+        (rejection) =>
+          expect(rejection.message).toEqual("events[0] must be an object")
       );
     });
 
-    it('should emit added events', async () => {
+    it("should emit added events", async () => {
       const emittedEvents: Event[] = [];
-      eventStore.onEvent(event => {
+      eventStore.onEvent((event) => {
         emittedEvents.push(event);
       });
       Object.assign(eventStore, {
         insertEventsInDb: () =>
-          Promise.resolve([createDbEvent({ id: 1 }), createDbEvent({ id: 2 })])
+          Promise.resolve([createDbEvent({ id: 1 }), createDbEvent({ id: 2 })]),
       });
 
       await eventStore.addAll([createEvent(), createEvent()]);
@@ -126,53 +128,53 @@ describe('Event store', () => {
     });
   });
 
-  describe('on find', () => {
-    it('should assert criteria are present', () => {
+  describe("on find", () => {
+    it("should assert criteria are present", () => {
       const find = () => eventStore.find(3 as any);
 
-      expect(find).toThrow('criteria must be an object');
+      expect(find).toThrow("criteria must be an object");
     });
 
-    it('should assert criteria are an object', () => {
+    it("should assert criteria are an object", () => {
       const find = () => eventStore.find(3 as any);
 
-      expect(find).toThrow('criteria must be an object');
+      expect(find).toThrow("criteria must be an object");
     });
 
-    it('should assert type is a string when defined', () => {
+    it("should assert type is a string when defined", () => {
       const find = () => eventStore.find({ type: 3 as any });
 
-      expect(find).toThrow('criteria#type must be a string');
+      expect(find).toThrow("criteria#type must be a string");
     });
 
-    it('should assert types are an array when defined', () => {
+    it("should assert types are an array when defined", () => {
       const find = () => eventStore.find({ types: 3 as any });
 
-      expect(find).toThrow('criteria#types must be an array');
+      expect(find).toThrow("criteria#types must be an array");
     });
 
-    it('should assert types are an array containing strings when defined', () => {
+    it("should assert types are an array containing strings when defined", () => {
       const find = () => eventStore.find({ types: [3] as any });
 
-      expect(find).toThrow('criteria#types[0] must be a string');
+      expect(find).toThrow("criteria#types[0] must be a string");
     });
 
-    it('should assert target id is a string when defined', () => {
+    it("should assert target id is a string when defined", () => {
       const find = () => eventStore.find({ targetId: 3 as any });
 
-      expect(find).toThrow('criteria#targetId must be a string');
+      expect(find).toThrow("criteria#targetId must be a string");
     });
 
-    it('should assert target type is a string when defined', () => {
+    it("should assert target type is a string when defined", () => {
       const find = () => eventStore.find({ targetType: 3 as any });
 
-      expect(find).toThrow('criteria#targetType must be a string');
+      expect(find).toThrow("criteria#targetType must be a string");
     });
 
-    it('should assert after id is a string when defined', () => {
+    it("should assert after id is a string when defined", () => {
       const find = () => eventStore.find({ afterId: 3 as any });
 
-      expect(find).toThrow('criteria#afterId must be a string');
+      expect(find).toThrow("criteria#afterId must be a string");
     });
   });
 });
