@@ -1,4 +1,4 @@
-export * from './samples';
+export * from "./samples";
 
 export const connectionString =
-  'postgres://postgres@localhost:5432/eventstoretests';
+  "postgres://postgres@localhost:5432/eventstoretests";
