@@ -1,3 +1,3 @@
-module.exports = Object.assign({}, require('./jest.config.js'), {
-  testRegex: '/lib/.*\\.e2e\\.ts$'
+module.exports = Object.assign({}, require("./jest.config.js"), {
+  testRegex: "/lib/.*\\.e2e\\.ts$",
 });
